Javascript, main.pyのリファクタリング
