+++ app/__init__.py
@@ -5,7 +5,7 @@ from app.models import User
 
 # Flask-Loginの設定
 login_manager = LoginManager()
-login_manager.login_view = 'main.login'
+login_manager.login_view = 'auth.login'
 login_manager.login_message = 'ログインしてください。'
 
 @login_manager.user_loader
@@ -15,18 +15,19 @@ def load_user(user_id):
 def create_app():
     app = Flask(__name__)
     app.config.from_object(Config)
-
     # Flask-Loginの初期化
     login_manager.init_app(app)
 
     # ブループリントのインポート
-
-    from app.routes import bp
-
-
+    from app.main import bp
+    from app.users import user_bp
+    from app.routes import auth_bp
+    from app.routes import data_management_bp
+    from app.routes import business_day_bp
     # ブループリントの登録
-
     app.register_blueprint(bp, url_prefix='/')
-
-
+    app.register_blueprint(user_bp, url_prefix="/user")
+    app.register_blueprint(auth_bp, url_prefix='/auth')
+    app.register_blueprint(data_management_bp)
+    app.register_blueprint(business_day_bp)
     return app
