+++ app/routes/__init__.py
@@ -1,0 +1,7 @@
+
+# app/routes/__init__.py
+from .auth import auth_bp
+
+# ブループリントの登録
+from .data_management import data_management_bp
+from .business_day import business_day_bp
