+++ app/routes/auth.py
@@ -1,0 +1,24 @@
+from flask import Blueprint, render_template, request, flash, redirect, url_for
+from flask_login import login_user, logout_user, login_required, current_user
+from app.models import User
+import logging
+
+auth_bp = Blueprint("auth", __name__, static_folder='static', template_folder='templates')
+@auth_bp.route('/login', methods=['GET', 'POST'])
+def login():
+    if request.method == 'POST':
+        username = request.form.get('username')
+        password = request.form.get('password')
+        user = User.authenticate(username, password)
+        if user:
+            login_user(user)
+            return redirect(url_for('main.index'))  
+        else:
+            flash('ログインに失敗しました。')
+    return render_template('login.html')
+
+@auth_bp.route('/logout')
+@login_required
+def logout():
+    logout_user()
+    return redirect(url_for('main.index'))
