+++ app/templates/index.html
@@ -10,7 +10,7 @@
 <body>
     <div class="background-overlay">
         <div class="page-container"> 
-            <h1>七曲署売上集計アプリ</h1> <a class="btn" href="/logout">ログアウト</a>
+            <h1>七曲署売上集計アプリ</h1> <a class="btn" href="auth/logout">ログアウト</a>
 
             <!-- Flash Messages -->
             {% with messages = get_flashed_messages(with_categories=true) %}
